+++ Content/FileHandling/IFileHandler.cs
@@ -10,7 +10,7 @@
  * Description = Interface for handling file upload and download
  *****************************************************************************/
 
-namespace Content
+namespace Content.FileHandling
 {
     /// <summary>
     /// Interface to upload and download files to and from server
@@ -19,17 +19,17 @@ namespace Content
     {
         List<string> _filesList { get; set; }
         /// <summary>
-        /// Upload a file from server
+        /// Upload a file to encoded as an XML
         /// </summary>
         /// <param name="filepath">path to file</param>
         /// <param name="sessionID">ID of this session</param>
         void Upload(string filepath, string sessionID);
 
         /// <summary>
-        /// Download a file from the server, pressumably queued for this session
+        /// Handle a received file by saving it to a directory
         /// </summary>
         /// <param name="sessionID">ID of this session</param>
         /// <returns></returns>
-        string Download(string sessionID);
+        void HandleRecieve(string sessionID, string encoding);
     }
 }
